chg:rename settings page spec; add tests for defaults + page reload test for dark mode
